+++ util-core/src/main/scala/com/twitter/io/BufReader.scala
@@ -2,6 +2,8 @@ package com.twitter.io
 
 import com.twitter.util.Future
 import java.util.NoSuchElementException
+import scala.annotation.tailrec
+import scala.collection.mutable.ListBuffer
 
 object BufReader {
 
@@ -49,4 +51,76 @@ object BufReader {
 
     loop(Buf.Empty)
   }
+
+  // see BufReader.chunked
+  private final class ChunkedFramer(chunkSize: Int) extends (Buf => Seq[Buf]) {
+    require(chunkSize > 0, s"chunkSize should be > 0 but was $chunkSize")
+
+    @tailrec
+    private def loop(acc: ListBuffer[Buf], in: Buf): Seq[Buf] = {
+      if (in.length < chunkSize) (acc :+ in).toSeq
+      else {
+        loop(
+          acc :+ in.slice(0, chunkSize),
+          in.slice(chunkSize, in.length)
+        )
+      }
+    }
+
+    def apply(in: Buf): Seq[Buf] = {
+      loop(ListBuffer(), in)
+    }
+  }
+
+  // see BufReader.framed
+  private final class Framed(r: Reader[Buf], framer: Buf => Seq[Buf])
+      extends Reader[Buf]
+      with (Option[Buf] => Future[Option[Buf]]) {
+
+    private[this] var frames: Seq[Buf] = Nil
+
+    // we only enter here when `frames` is empty.
+    def apply(in: Option[Buf]): Future[Option[Buf]] = synchronized {
+      in match {
+        case Some(data) =>
+          frames = framer(data)
+          read()
+        case None =>
+          Future.None
+      }
+    }
+
+    def read(): Future[Option[Buf]] = synchronized {
+      if (frames.isEmpty) {
+        // flatMap to `this` to prevent allocating
+        r.read().flatMap(this)
+      } else {
+        val nextFrame = frames.head
+        frames = frames.tail
+        Future.value(Some(nextFrame))
+      }
+    }
+
+    def discard(): Unit = synchronized {
+      frames = Seq.empty
+      r.discard()
+    }
+
+    def onClose: Future[StreamTermination] = r.onClose
+  }
+
+  /**
+   * Chunk the output of a given [[Reader]] by at most `chunkSize` (bytes). This consumes the
+   * reader.
+   */
+  def chunked(r: Reader[Buf], chunkSize: Int): Reader[Buf] =
+    new Framed(r, new ChunkedFramer(chunkSize))
+
+  /**
+   * Wraps a [[ Reader[Buf] ]] and emits frames as decided by `framer`.
+   *
+   * @note The returned `Reader` may not be thread safe depending on the behavior
+   *       of the framer.
+   */
+  def framed(r: Reader[Buf], framer: Buf => Seq[Buf]): Reader[Buf] = new Framed(r, framer)
 }